Refactor Actions
