+++ actions/city-weather.js
@@ -5,8 +5,6 @@ export const ADD_CITY_FAILURE = 'ADD_CITY_FAILURE'
 export const REMOVE_CITY = 'REMOVE_CITY'
 export const FETCH_CITY_WEATHER_SUCCESS = 'FETCH_CITY_WEATHER_SUCCESS'
 export const FETCH_CITY_WEATHER_FAILURE = 'FETCH_CITY_WEATHER_FAILURE'
-export const UPDATE_UNITS_SUCCESS = 'UPDATE_UNITS_SUCCESS'
-// ACTIONS
 
 export const addCity = (city) => async dispatch => {
     try {
@@ -18,11 +16,9 @@ export const addCity = (city) => async dispatch => {
     }
     
 }
-
 export const removeCity = (data) => dispatch => {
     dispatch({ type: REMOVE_CITY, data })
 }
-
 export const fetchCityWeather = (id) => async dispatch => {
     try {
         const current = await getCityWeather(id)
@@ -33,8 +29,4 @@ export const fetchCityWeather = (id) => async dispatch => {
         console.log(err.msg)
         dispatch({ type: FETCH_CITY_WEATHER_FAILURE, err })
     }
-}
-
-export const updateUnits = (units) => dispatch => {
-    dispatch({ type: UPDATE_UNITS_SUCCESS, units })
 }
