+++ actions/units.js
@@ -1,0 +1,5 @@
+export const UPDATE_UNITS_SUCCESS = 'UPDATE_UNITS_SUCCESS'
+
+export const updateUnits = (units) => dispatch => {
+    dispatch({ type: UPDATE_UNITS_SUCCESS, units })
+}
