+++ api/weather.js
@@ -1,6 +1,6 @@
 import axios from 'axios'
 
-const APP_ID = '7ee2d8e6706dc28e3c12e760a6d04d9f'
+const APP_ID = '7ee2d8e6706dc28e3c12e760a6d04d9f' //FIX move to ENV later
 const api = axios.create({
     baseURL: `http://api.openweathermap.org/data/2.5/`
 })
