+++ components/location-search-input.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import PlacesAutocomplete from 'react-places-autocomplete'
-import { addCity } from '../actions/'
+import { addCity } from '../actions/city-weather'
 import { InputBox, AutocompleteWrap, AutocompleteItems, AutocompleteItemsContainner } from '../components/styled/autocomplete'
 
 class LocationSearchInput extends React.Component {
